task 7 done

+++ 7/7.go
@@ -9,12 +9,12 @@ func main() {
 	fmt.Println(reverse(-123))
 }
 
-func reverse(num int) int {
+func reverse(num int) int32 {
 	res := 0
 	for num != 0 {
 		remainder := num % 10
 		res = (res * 10) + remainder
 		num /= 10
 	}
-	return res
+	return int32(res)
 }
